conversion util program
